requirements.txt
# Project H.I.V.E Main Requirements

# Web frontend
Flask
python-dotenv
requests

# API servers
fastapi
uvicorn
pydantic

# Configuration and data processing
PyYAML
aiohttp

# Windows Admin Auth
pywin32; platform_system == "Windows"

# Linux/macOS PAM Auth
simplepam; platform_system != "Windows"